Upgrading libtiff: Patches to libtiff 4.0.3 for Qt usage
This is the collection of those earlier patches to libtiff in Qt that
are still relevant, updated as required. For details, refer to the git
history of the src/3rdparty/libtiff directory of Qt 4 and Qt
5/qtimageformats.

All defines which happened in tiffconf.h were commented in tif_config.h.

Done-with: aavit <eirik.aavitsland@digia.com>
Task-number: QTBUG-25409
Change-Id: I8dc43f0dfe47e6472c3475d9e60ee304dc54ee3b
Reviewed-by: aavit <eirik.aavitsland@digia.com>