Build MNG and Jpeg2000 plugins only if system libs found
Upstream development libmng and libjasper is mostly stalled,
so security issues are dealt with by the linux distributions
instead. Hence, the stale bundled copy of the 3rd party source code of
those libraries can not be generally recommended.

Leave the bundled versions in the repo for now as a transitional
measure, so that the bundled-lib plugins still can be built manually
(by doing qmake & make in the relevant src/plugins/imageformats
subdirectory). In a later Qt version, the bundled copies of those libs
should be removed.

Change-Id: I651417d91ef0f43cb6abe638f97968576a78e2f0
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>
Reviewed-by: Lars Knoll <lars.knoll@theqtcompany.com>