Skip (pass) each autotest if the corresponding handler not present
Add a check at initialization of each autotest, skipping it if the
relevant plugin is not installed. This is the same as in the
qtimagereader autotest in qtbase (SKIP_IF_UNSUPPORTED macro).

Reason for change: Not all format handlers are supported on all
platforms. Hence, the suite of autotests ("make check") should not
fail on account of a missing plugin.

Change-Id: Ic0f4da68a9527831f017db0f326afe6e631efd9e
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@theqtcompany.com>
Reviewed-by: Liang Qi <liang.qi@theqtcompany.com>