QTgaFile: fix Coverity warning re:mixing enum types
Coverity complained that the 'bytes' variable was first compared to
a QTgaFile::HeaderOffset, then, later, to a QTgaFile::FooterOffset.

This is of course a false positive, but one that's trivial to fix,
by not using a variable in the first place.

Adjust to Qt coding standard as a drive-by.

Coverity-Id: 22048
Change-Id: If1a45aa5b0c8ea23cab2cefeccb2f1dfe5f03375
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
(cherry picked from commit 33b9fc4e062537a5ab8d4c60c72573f868ab3f13)