QTgaFile: fix parsing of TGA16 rgb data
The code tries to expand a 16-bit value of the form 0bABBBBBGGGGGRRRRR
into a 32-bit QRgb, but got the operator precedence wrong:

  << has higher precedence than binary &

This made the first operand of the |-chain (BBBBB) unconditionally
zero.  The second operand had the same precedence problem, but didn't
decay into a tautological value like the first one did.

Fix by adding another set of parentheses.

The test coverage for this security-relevant piece of code is quite
obviously insufficient, and should be increased, or else the format be
dropped.

[ChangeLog][TGA] Fixed reading of TGA-16 formats.

Coverity-Id: 21782
Change-Id: I7019be8fe22e480c40192e0c1916b1d2bebf71cc
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
Reviewed-by: Giuseppe D'Angelo <giuseppe.dangelo@kdab.com>