Fix tiffhandler: ensure saved tiffs have suitable ROWSPERSTRIP tag set
When writing an image without compression, libtiff will automatically
store it in strips of default size, and add a ROWSPERSTRIP tag
accordingly. However, if compression is enabled, libtiff does not do
this. The lack of this tag may create trouble for readers.
Fix by explicitly setting the tag in all cases.

[ChangeLog][TIFF] Ensure saved tiffs have suitable ROWSPERSTRIP tag set

Task-number: QTBUG-68609
Change-Id: I838c83be25158d7f13e220098e960010b8cb2789
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>