Fix and re-work compile tests
The compile tests were seriously broken, and we didn't notice, because
configure leniently ignores the second level "test" object and the
inline sources provide the absolute path.

Merge the 2nd level test objects with the right ones.

Remove superfluous label entries.

Remove superfluous qmake entries. The inline sources already provide
the right -l argument.

Remove superfluous return statements. The generated main function
already returns 0.

Remove superfluous "type" entry, because "compile" is the default anyways.

Note that the jasper version test is bogus, because these are compile
tests that are not actually executed. The reference to JAS_VERSION is
kept to test for existence of this define.

Fixes: QTBUG-78473
Change-Id: I1d4fa634e29329cfc34c0a2f78b2eb730ef84d50
Reviewed-by: Liang Qi <liang.qi@qt.io>