Update commercial license headers
Updated header.COMM to the files in tqtc-qtimageformats.
Documentation files, examples or tests are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the line count causes failures in tests.

Task-number: QTQAINFRA-4165
Change-Id: I610d62cee803b55d3676faa08d8947c6a8201355
Reviewed-by: Jani Heikkinen <jani.heikkinen@qt.io>