Fix mng image plugin in cmake build
Add cmake/FindLibmng.cmake based on qtbase/cmake/FindLibb2.cmake .
Enable mng if found and link to the imported target provided by
FindLibmng.cmake so that mng can be used.
Libmng 1 did not come with a pkgconfig file and it is still
in use so FindLibmng will use the normal cmake procedures if
pkgconfig fails to find libmng.

Pick-to: 6.1
Change-Id: Iecf4ede700b1bfdab84c98c7333547f0bcecc6b3
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>