Update commercial license headers
Updated header.COMM to the files in tqtc-qtimageformats.
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the line count causes failures in tests.

Task-number: QTQAINFRA-4945
Change-Id: Ie3652c8c5eadf600977d05f6897076c91ee53f7e
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>