Update HEIF test image
The original image stopped working on macOS 13 Ventura, both with our
image plugin and in Preview. It's unclear why the original image works
fine in Preview on macOS 12, but that's an Apple bug, and our image
plugin tests should not have to worry about platform differences like
that.

Task-number: QTBUG-104210
Pick-to: 6.4 6.2
Change-Id: If57fce379a2d9bcb92757e3f2f7e903b079e700b
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>